Final simplification
